+++ sockets/sockets.h
@@ -54,12 +54,12 @@ int connectSocket(char *address, char *port);
     Receive the file descriptor of the socket, a pointer to where to store the data and the maximum size avaliable
     Returns 1 on successful receipt, or 0 if the connection has finished
 */
-int recvString(int connection_fd, void *buffer, int size);
+int recvString(int connection_fd, void *buffer, size_t size);
 
 /*
     Send a message with error validation
     Receive the file descriptor, the pointer to the data, and the size of the data to send
 */
-void sendString(int connection_fd, void *buffer, int size);
+void sendString(int connection_fd, void *buffer, size_t size);
 
 #endif
